feat: signing api (#13)
* Allow coordinator to be set manually

* Add manual coordinator test

* Implement across message handler

* Send signature to all relayers

* Implement signature cache to store generated signatures

* Speed up process coordination

* Implement signature cache to store  generated signatures

* Add signature cache tests

* Allow for coordinator to notify other relayers of an across message

* Reduce tss timeout

* Implement calculation of the unlock hash that matches on-chain implementation

* Add across message handler tests

* Update mocks to maintained version

* Fix test race condition

* Convert abi to constant

* Use source chain id property from across data

* Use message destination as across deposit source

* Use coordinator from the host in message handler

* Send across message result over the err chanel

* Return api errors as json

* Implement api serve function

* Add signing handler tests

* Ignore valid json encode errors

* Remove print

* Bump cache action

* Use a custom big.Int type that decode big int from string

* Add read timeout

* Add custom lint configuration

* Remove messy linters

* Lint

* Remove extra nolint

* Remove linter errors

+++ comm/subID.go
@@ -16,7 +16,7 @@ import (
 type SubscriptionID string
 
 func NewSubscriptionID(sessionID string, msgType MessageType) SubscriptionID {
-	return SubscriptionID(fmt.Sprintf("%s-%d-%d", sessionID, msgType, uint32(time.Now().UnixNano())))
+	return SubscriptionID(fmt.Sprintf("%s-%d-%d", sessionID, msgType, time.Now().UnixNano()))
 }
 
 func (sID SubscriptionID) SessionID() string {
